Move refactor EmployeePage in employee directory
